requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
aiofiles
python-dotenv
deepface
opencv-python-headless